backend/requirements.txt
django==5.2.11
djangorestframework==3.16.1
gunicorn==25.0.1
mysqlclient==2.2.7
python-dotenv==1.2.1
django-storages==1.14.6
boto3==1.42.40
django-cors-headers==4.9.0
pymysql==1.1.0